Views controller splited to Users and Moderators controllers
